# perf(s2n-quic-core): optimize varint encoding