… quic/s2n-quic-core/src/varint/snapshots/s2n_quic_core__varint__tests__table_2_pow_0_.snap
… quic/s2n-quic-core/src/varint/snapshots/s2n_quic_core__varint__tests__table_2_pow_13_.snap
… quic/s2n-quic-core/src/varint/snapshots/s2n_quic_core__varint__tests__table_2_pow_14_.snap
… quic/s2n-quic-core/src/varint/snapshots/s2n_quic_core__varint__tests__table_2_pow_1_.snap
… quic/s2n-quic-core/src/varint/snapshots/s2n_quic_core__varint__tests__table_2_pow_29_.snap
… quic/s2n-quic-core/src/varint/snapshots/s2n_quic_core__varint__tests__table_2_pow_30_.snap
… quic/s2n-quic-core/src/varint/snapshots/s2n_quic_core__varint__tests__table_2_pow_5_.snap
… quic/s2n-quic-core/src/varint/snapshots/s2n_quic_core__varint__tests__table_2_pow_61_.snap
… quic/s2n-quic-core/src/varint/snapshots/s2n_quic_core__varint__tests__table_2_pow_6_.snap