### Description of changes: 

This PR focuses on improving varint encoding a bit by:

* improving the table function to be completely branchless (see https://godbolt.org/z/9xrxd1osd)
* Adding a special case when we have at least 8 bytes in the target buffer and just doing a single 64 bit `mov`, rather than having to branch based on our integer size:
  https://github.com/aws/s2n-quic/blob/b84785290f147a8b97616d449ab93632c8842bab/quic/s2n-quic-core/src/varint/table.rs#L262-L271

I've also cleaned the code up a bit by moving the table functions into a separate module with an named struct, rather than a `(u64, usize, u64)`.

Performance has improved by about 30-40%

```
varint/encode/array/1   time:   [1.6209 ns 1.6211 ns 1.6212 ns]
                        change: [-40.046% -39.123% -37.678%] (p = 0.00 < 0.05)
varint/encode/slice/1   time:   [2.4321 ns 2.4329 ns 2.4341 ns]
                        change: [-33.393% -33.335% -33.266%] (p = 0.00 < 0.05)
varint/encode/array_16/1
                        time:   [38.505 ns 38.515 ns 38.529 ns]
                        change: [-31.424% -31.257% -31.005%] (p = 0.00 < 0.05)
varint/encode/slice_16/1
                        time:   [38.676 ns 38.790 ns 38.889 ns]
                        change: [-38.523% -38.430% -38.354%] (p = 0.00 < 0.05)
```

### Testing:

The existing tests should prevent any regressions in behavior.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.