@WesleyRosenblum 2023-10-20
Can you summarize (maybe in the PR description) how the optimization works and why it achieves 30-40% better perf?